intf: check the volume to verify that VLC is still running
On most VLC versions, closing VLC should stop the modules. However,
it seems that in some cases the VLC process is kept running until
the modules stop. In those latter cases, TraktForVLC was causing
VLC to hang at the end of the execution. Checking the volume will
thus allow to see if VLC is still opened, and stop properly if it
is not.

Fixes #100

Signed-off-by: Raphaël Beamonte <raphael.beamonte@gmail.com>